requirements.txt
python-dotenv
streamlit
streamlit-options-menu
google-genai